// src/generated/models/Cr871_assessmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr871_assessmentscr871_finaldecision = {
  144610000: 'Approved',
  144610001: 'ApprovedwithConditions',
  144610002: 'Deferred',
  144610003: 'Rejected'
} as const;
export type Cr871_assessmentscr871_finaldecision = keyof typeof Cr871_assessmentscr871_finaldecision;
export const Cr871_assessmentscr871_riskband = {
  144610000: 'LowRisk',
  144610001: 'MediumRisk',
  144610002: 'HighRisk',
  144610003: 'CriticalRisk'
} as const;
export type Cr871_assessmentscr871_riskband = keyof typeof Cr871_assessmentscr871_riskband;
export const Cr871_assessmentscr871_riskrating = {
  144610000: 'Critical',
  144610001: 'High',
  144610002: 'Medium',
  144610003: 'Low'
} as const;
export type Cr871_assessmentscr871_riskrating = keyof typeof Cr871_assessmentscr871_riskrating;
export const Cr871_assessmentscr871_status = {
  144610000: 'Invited',
  144610001: 'InProgress',
  144610002: 'Submitted',
  144610003: 'UnderReview',
  144610004: 'Complete',
  144610005: 'Lapsed'
} as const;
export type Cr871_assessmentscr871_status = keyof typeof Cr871_assessmentscr871_status;
export const Cr871_assessmentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr871_assessmentsstatecode = keyof typeof Cr871_assessmentsstatecode;
export const Cr871_assessmentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr871_assessmentsstatuscode = keyof typeof Cr871_assessmentsstatuscode;

export interface Cr871_assessmentsBase {
  cr871_assessmentid: string;
  cr871_assessmentid1?: string;
  cr871_assessoremail?: string;
  cr871_assessornotes?: string;
  cr871_duedate?: string;
  cr871_finaldecision?: Cr871_assessmentscr871_finaldecision;
  cr871_invitedate?: string;
  cr871_overallscore?: number;
  cr871_remindercount?: number;
  cr871_riskband?: Cr871_assessmentscr871_riskband;
  cr871_riskrating: Cr871_assessmentscr871_riskrating;
  cr871_scoredelta?: number;
  cr871_status?: Cr871_assessmentscr871_status;
  cr871_submitdate?: string;
  cr871_vendorcontactemail?: string;
  "cr871_VendorID@odata.bind"?: string;
  cr871_vendorname?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr871_assessmentsstatecode;
  statuscode?: Cr871_assessmentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr871_assessments extends Cr871_assessmentsBase {
  cr871_finaldecisionname?: string;
  cr871_riskbandname?: string;
  cr871_riskratingname?: string;
  cr871_statusname?: string;
  cr871_vendoridname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr871_vendorid?: object;
  _cr871_vendorid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}